Data.Array.Accelerate.Algebra.IntegralDomain

Description

Documentation

divModZero :: forall a. (C (Exp a), C (Exp a), Elt a) => Exp a -> Exp a -> Exp (a, a)

Allows division by zero. If the denominator is zero, the numerator is returned as the remainder.

divides :: (C (Exp a), C (Exp a), Elt a) => Exp a -> Exp a -> Exp Bool

Test if the first argument evenly divides the second

sameResidueClass :: (C (Exp a), C (Exp a), Elt a) => Exp a -> Exp a -> Exp a -> Exp Bool

even :: (C (Exp a), C (Exp a), Elt a) => Exp a -> Exp Bool

odd :: (C (Exp a), C (Exp a), Elt a) => Exp a -> Exp Bool

Orphan instances

C (Exp Int)

Methods

div :: Exp Int -> Exp Int -> Exp Int

mod :: Exp Int -> Exp Int -> Exp Int

divMod :: Exp Int -> Exp Int -> (Exp Int, Exp Int)

C (Exp Int8)

Methods

div :: Exp Int8 -> Exp Int8 -> Exp Int8

mod :: Exp Int8 -> Exp Int8 -> Exp Int8

divMod :: Exp Int8 -> Exp Int8 -> (Exp Int8, Exp Int8)

C (Exp Int16)

Methods

div :: Exp Int16 -> Exp Int16 -> Exp Int16

mod :: Exp Int16 -> Exp Int16 -> Exp Int16

divMod :: Exp Int16 -> Exp Int16 -> (Exp Int16, Exp Int16)

C (Exp Int32)

Methods

div :: Exp Int32 -> Exp Int32 -> Exp Int32

mod :: Exp Int32 -> Exp Int32 -> Exp Int32

divMod :: Exp Int32 -> Exp Int32 -> (Exp Int32, Exp Int32)

C (Exp Int64)

Methods

div :: Exp Int64 -> Exp Int64 -> Exp Int64

mod :: Exp Int64 -> Exp Int64 -> Exp Int64

divMod :: Exp Int64 -> Exp Int64 -> (Exp Int64, Exp Int64)

C (Exp Word)

Methods

div :: Exp Word -> Exp Word -> Exp Word

mod :: Exp Word -> Exp Word -> Exp Word

divMod :: Exp Word -> Exp Word -> (Exp Word, Exp Word)

C (Exp Word8)

Methods

div :: Exp Word8 -> Exp Word8 -> Exp Word8

mod :: Exp Word8 -> Exp Word8 -> Exp Word8

divMod :: Exp Word8 -> Exp Word8 -> (Exp Word8, Exp Word8)

C (Exp Word16)

Methods

div :: Exp Word16 -> Exp Word16 -> Exp Word16

mod :: Exp Word16 -> Exp Word16 -> Exp Word16

divMod :: Exp Word16 -> Exp Word16 -> (Exp Word16, Exp Word16)

C (Exp Word32)

Methods

div :: Exp Word32 -> Exp Word32 -> Exp Word32

mod :: Exp Word32 -> Exp Word32 -> Exp Word32

divMod :: Exp Word32 -> Exp Word32 -> (Exp Word32, Exp Word32)

C (Exp Word64)

Methods

div :: Exp Word64 -> Exp Word64 -> Exp Word64

mod :: Exp Word64 -> Exp Word64 -> Exp Word64

divMod :: Exp Word64 -> Exp Word64 -> (Exp Word64, Exp Word64)

(C (Exp a), C (Exp b), Elt a, Elt b) => C (Exp (a, b))

Methods

div :: Exp (a, b) -> Exp (a, b) -> Exp (a, b)

mod :: Exp (a, b) -> Exp (a, b) -> Exp (a, b)

divMod :: Exp (a, b) -> Exp (a, b) -> (Exp (a, b), Exp (a, b))