Data.Array.Accelerate.Algebra.PrincipalIdealDomain

Description

Documentation

class (C a, C a) => C a where

Minimal complete definition

Nothing

Methods

extendedGCD :: a -> a -> (a, (a, a))

gcd :: a -> a -> a

lcm :: a -> a -> a

coprime :: (C (Exp a), C (Exp a)) => Exp a -> Exp a -> Exp Bool

Two integers a and b are coprime (aka: relatively prime, mutually prime) if the only positive integer that divides them both is one.

euclid :: forall a. (C (Exp a), C (Exp a), Elt a) => (Exp a -> Exp a -> Exp a) -> Exp a -> Exp a -> Exp a

An efficient method for computing the greatest common divisor of two numbers.

extendedEuclid :: forall a. (C (Exp a), C (Exp a), C (Exp a), C (Exp a), Elt a) => (Exp a -> Exp a -> Exp a) -> Exp a -> Exp a -> Exp (a, (a, a))

Implementation of the extended Euclid algorithm which computes, besides the greatest common divisor of integers a and b, the coefficients of Bézout's identity x and y such that:

ax + by = gcd(a, b)

https://en.wikipedia.org/wiki/Extended_Euclidean_algorithm#Pseudocode