Data.Array.Accelerate.Algebra.Units

Description

Documentation

class C a => C a where

This class lets us deal with units in a ring. isUnit tells us whether an element is a unit, while the other operations let us canonically write an element as a unit times another element.

This instance shadows C with an instance of isUnit compatible with Accelerate.

Minimal complete definition

isUnit, (stdUnit | stdUnitInv)

Methods

isUnit :: a -> Exp Bool

stdAssociate :: a -> a

stdUnit :: a -> a

stdUnitInv :: a -> a