Data.Array.Accelerate.Algebra.IntegralDomain

Description

Documentation

divModZero :: forall a. (C (Exp a), C (Exp a), Elt a) => Exp a -> Exp a -> Exp (a, a)

Allows division by zero. If the denominator is zero, the numerator is returned as the remainder.

divides :: (C (Exp a), C (Exp a), Elt a) => Exp a -> Exp a -> Exp Bool

Test if the first argument evenly divides the second

sameResidueClass :: (C (Exp a), C (Exp a), Elt a) => Exp a -> Exp a -> Exp a -> Exp Bool

even :: (C (Exp a), C (Exp a), Elt a) => Exp a -> Exp Bool

odd :: (C (Exp a), C (Exp a), Elt a) => Exp a -> Exp Bool