Data.Array.Accelerate.Algebra.PrincipalIdealDomain

Description

Documentation

class (C a, C a) => C a where

Methods

extendedGCD :: a -> a -> (a, (a, a))

gcd :: a -> a -> a

lcm :: a -> a -> a

coprime :: (C (Exp a), C (Exp a)) => Exp a -> Exp a -> Exp Bool

Two integers a and b are coprime (aka: relatively prime, mutually prime) if the only positive integer that divides them both is one.

euclid :: forall a. (C (Exp a), C (Exp a), Elt a) => (Exp a -> Exp a -> Exp a) -> Exp a -> Exp a -> Exp a

An efficient method for computing the greatest common divisor of two numbers.

extendedEuclid :: forall a. (C (Exp a), C (Exp a), C (Exp a), C (Exp a), Elt a) => (Exp a -> Exp a -> Exp a) -> Exp a -> Exp a -> Exp (a, (a, a))

Implementation of the extended Euclid algorithm which computes, besides the greatest common divisor of integers a and b, the coefficients of Bézout's identity x and y such that:

ax + by = gcd(a, b)

https://en.wikipedia.org/wiki/Extended_Euclidean_algorithm#Pseudocode

Orphan instances

C (Exp Int)

Methods

extendedGCD :: Exp Int -> Exp Int -> (Exp Int, (Exp Int, Exp Int))

gcd :: Exp Int -> Exp Int -> Exp Int

lcm :: Exp Int -> Exp Int -> Exp Int

C (Exp Int8)

Methods

extendedGCD :: Exp Int8 -> Exp Int8 -> (Exp Int8, (Exp Int8, Exp Int8))

gcd :: Exp Int8 -> Exp Int8 -> Exp Int8

lcm :: Exp Int8 -> Exp Int8 -> Exp Int8

C (Exp Int16)

Methods

extendedGCD :: Exp Int16 -> Exp Int16 -> (Exp Int16, (Exp Int16, Exp Int16))

gcd :: Exp Int16 -> Exp Int16 -> Exp Int16

lcm :: Exp Int16 -> Exp Int16 -> Exp Int16

C (Exp Int32)

Methods

extendedGCD :: Exp Int32 -> Exp Int32 -> (Exp Int32, (Exp Int32, Exp Int32))

gcd :: Exp Int32 -> Exp Int32 -> Exp Int32

lcm :: Exp Int32 -> Exp Int32 -> Exp Int32

C (Exp Int64)

Methods

extendedGCD :: Exp Int64 -> Exp Int64 -> (Exp Int64, (Exp Int64, Exp Int64))

gcd :: Exp Int64 -> Exp Int64 -> Exp Int64

lcm :: Exp Int64 -> Exp Int64 -> Exp Int64