Data.Array.Accelerate.Algebra.Units

Description

Documentation

class C a => C a where

This class lets us deal with units in a ring. isUnit tells us whether an element is a unit, while the other operations let us canonically write an element as a unit times another element.

This instance shadows C with an instance of isUnit compatible with Accelerate.

Minimal complete definition

isUnit, (stdUnit | stdUnitInv)

Methods

isUnit :: a -> Exp Bool

stdAssociate :: a -> a

stdUnit :: a -> a

stdUnitInv :: a -> a

Orphan instances

C (Exp Int)

Methods

isUnit :: Exp Int -> Bool

stdAssociate :: Exp Int -> Exp Int

stdUnit :: Exp Int -> Exp Int

stdUnitInv :: Exp Int -> Exp Int

C (Exp Int8)

Methods

isUnit :: Exp Int8 -> Bool

stdAssociate :: Exp Int8 -> Exp Int8

stdUnit :: Exp Int8 -> Exp Int8

stdUnitInv :: Exp Int8 -> Exp Int8

C (Exp Int16)

Methods

isUnit :: Exp Int16 -> Bool

stdAssociate :: Exp Int16 -> Exp Int16

stdUnit :: Exp Int16 -> Exp Int16

stdUnitInv :: Exp Int16 -> Exp Int16

C (Exp Int32)

Methods

isUnit :: Exp Int32 -> Bool

stdAssociate :: Exp Int32 -> Exp Int32

stdUnit :: Exp Int32 -> Exp Int32

stdUnitInv :: Exp Int32 -> Exp Int32

C (Exp Int64)

Methods

isUnit :: Exp Int64 -> Bool

stdAssociate :: Exp Int64 -> Exp Int64

stdUnit :: Exp Int64 -> Exp Int64

stdUnitInv :: Exp Int64 -> Exp Int64