Data.Array.Accelerate.Number.Complex

Description

Complex numbers

Cartesian form

data T a :: * -> *

real :: Elt a => Exp (T a) -> Exp a

Real part

imag :: Elt a => Exp (T a) -> Exp a

Imaginary part

(+:) :: Elt a => Exp a -> Exp a -> Exp (T a) infix 6

Construct a complex number from the real and imaginary parts

(-:) :: forall a. (Elt a, C (Exp a)) => Exp a -> Exp a -> Exp (T a)

Construct a complex number with negated imaginary part

scale :: forall a. (C (Exp a), Elt a) => Exp a -> Exp (T a) -> Exp (T a)

Scale a complex number by a real number.

exp :: forall a. (C (Exp a), Elt a) => Exp (T a) -> Exp (T a)

Exponential of a complex number with minimal type class constraints.

quarterLeft :: forall a. (C (Exp a), Elt a) => Exp (T a) -> Exp (T a)

Turn the point one quarter to the left

quarterRight :: forall a. (C (Exp a), Elt a) => Exp (T a) -> Exp (T a)

Turn the point one quarter to the right.

Polar form

toPolar :: (C (Exp a), C (Exp a), Elt a) => Exp (T a) -> Exp (a, a)

The function toPolar takes a complex number and returns a (magnitude, phase) pair in canonical form: the magnitude is non-negative, and the phase in the range (-pi, pi]; if the magnitude is zero, then so is the phase.

fromPolar :: (C (Exp a), Elt a) => Exp a -> Exp a -> Exp (T a)

Form a complex number from polar components of magnitude and phase.

cis :: forall a. (C (Exp a), Elt a) => Exp a -> Exp (T a)

cis t is a complex value with magnitude 1 and phase t (modulo 2*pi).

signum :: (C (Exp a), C (Exp a), Elt a) => Exp (T a) -> Exp (T a)

Scale a complex number to magnitude 1.

For a complex number z, abs z is a number with the magnitude of z, but oriented in the positive real direction, whereas signum z has the phase of z, but unit magnitude.

magnitude :: (C (Exp a), Elt a) => Exp (T a) -> Exp a

The non-negative magnitude of a complex number. This implementation respects the limited range of floating point numbers. The trivial implementation magnitude would overflow for floating point exponents greater than the half of the maximum admissible exponent.

floatMagnitude :: (P.RealFloat a, Algebraic.C a) => T a -> a floatMagnitude (Cons x y) = let k = max (P.exponent x) (P.exponent y) mk = - k in P.scaleFloat k (sqrt (P.scaleFloat mk x ^ 2 + P.scaleFloat mk y ^ 2))

The non-negative magnitude of a complex number.

magnitudeSqr :: forall a. (C (Exp a), Elt a) => Exp (T a) -> Exp a

Like NormedEuc.normSqr with lifted class constraints

phase :: (C (Exp a), C (Exp a), Elt a) => Exp (T a) -> Exp a

The phase of a complex number, in the range (-pi, pi]. If the magnitude is zero, then so is the phase.

Conjugate

conjugate :: forall a. (C (Exp a), Elt a) => Exp (T a) -> Exp (T a)

The conjugate of a complex number.

Orphan instances

cst a => IsProduct cst (T a)

Associated Types

type ProdRepr (T a) :: *

Methods

fromProd :: proxy cst -> T a -> ProdRepr (T a)

toProd :: proxy cst -> ProdRepr (T a) -> T a

prod :: proxy cst -> T a -> ProdR cst (ProdRepr (T a))

(Lift Exp a, Elt (Plain a)) => Lift Exp (T a)

Associated Types

type Plain (T a) :: *

Methods

lift :: T a -> Exp (Plain (T a))

Elt a => Unlift Exp (T (Exp a))

Methods

unlift :: Exp (Plain (T (Exp a))) -> T (Exp a)

Elt a => Elt (T a)

Methods

eltType :: T a -> TupleType (EltRepr (T a))

fromElt :: T a -> EltRepr (T a)

toElt :: EltRepr (T a) -> T a

(C (Exp a), Elt a) => C (Exp (T a))

Methods

zero :: Exp (T a)

(+) :: Exp (T a) -> Exp (T a) -> Exp (T a)

(-) :: Exp (T a) -> Exp (T a) -> Exp (T a)

negate :: Exp (T a) -> Exp (T a)

(C (Exp a), Elt a) => C (Exp (T a))

Methods

(*) :: Exp (T a) -> Exp (T a) -> Exp (T a)

one :: Exp (T a)

fromInteger :: Integer -> Exp (T a)

(^) :: Exp (T a) -> Integer -> Exp (T a)

(C (Exp a), Elt a) => C (Exp (T a))

Methods

(/) :: Exp (T a) -> Exp (T a) -> Exp (T a)

recip :: Exp (T a) -> Exp (T a)

fromRational' :: Rational -> Exp (T a)

(^-) :: Exp (T a) -> Integer -> Exp (T a)

(C (Exp a), C (Exp a), C (Exp a), Elt a) => C (Exp (T a))

Methods

abs :: Exp (T a) -> Exp (T a)

signum :: Exp (T a) -> Exp (T a)

(C (Exp a), C (Exp a), C (Exp a), C (Exp a), C (Exp a), C (Exp a), Ord a, Elt a) => C (Exp (T a))

Methods

sqrt :: Exp (T a) -> Exp (T a)

root :: Integer -> Exp (T a) -> Exp (T a)

(^/) :: Exp (T a) -> Rational -> Exp (T a)

(C (Exp a), C (Exp a), C (Exp a), C (Exp a), Ord a, Elt a) => C (Exp (T a))

Methods

pi :: Exp (T a)

exp :: Exp (T a) -> Exp (T a)

log :: Exp (T a) -> Exp (T a)

logBase :: Exp (T a) -> Exp (T a) -> Exp (T a)

(**) :: Exp (T a) -> Exp (T a) -> Exp (T a)

sin :: Exp (T a) -> Exp (T a)

cos :: Exp (T a) -> Exp (T a)

tan :: Exp (T a) -> Exp (T a)

asin :: Exp (T a) -> Exp (T a)

acos :: Exp (T a) -> Exp (T a)

atan :: Exp (T a) -> Exp (T a)

sinh :: Exp (T a) -> Exp (T a)

cosh :: Exp (T a) -> Exp (T a)

tanh :: Exp (T a) -> Exp (T a)

asinh :: Exp (T a) -> Exp (T a)

acosh :: Exp (T a) -> Exp (T a)

atanh :: Exp (T a) -> Exp (T a)

(C (Exp a), Elt a) => C (Exp (T a))

Methods

isZero :: Exp (T a) -> Exp Bool